requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
greenlet==1.1.3.post0
idna==3.4
inflect==6.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
protobuf==3.20.1
pydantic==1.10.2
PyPDF2==2.11.1
pytz==2022.5
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlacodegen==2.3.0
SQLAlchemy==1.4.42
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
